Create Dynamic Connectivity
To write a program to filter out extraneous pairs (pairs where both objects are in the same equivalence class) from the sequence. 
In other words, when the program reads a pair p and q from the input, it should write the pair to the output only if the pairs it has seen to that point do not imply that p is connected to q.
If the previous pairs do imply that p is connected to q then the program should ignore the pair p and q then proceed to read in the next pair.